feat: add torus equivalent value to percent and revert icon change
